+++ lib/pages/Inventory/components/weapon_display.dart
@@ -14,10 +14,12 @@ class WeaponDisplay extends ConsumerWidget {
   final bool isQuickSelect;
   final String damage;
   final String name;
+  final bool canDelete;
   WeaponDisplay({
     super.key,
     this.damage = '',
     this.isQuickSelect = false,
+    this.canDelete = true,
     required this.guid,
     required this.name,
   });
@@ -120,7 +122,9 @@ class WeaponDisplay extends ConsumerWidget {
                                 topRight: Radius.circular(10),
                                 bottomRight: Radius.circular(10))),
                         child: Padding(
-                          padding: const EdgeInsets.only(left: 6.0),
+                          padding: canDelete
+                              ? const EdgeInsets.only(left: 6.0)
+                              : const EdgeInsets.fromLTRB(6, 12, 0, 12),
                           child: Row(
                             mainAxisAlignment: MainAxisAlignment.spaceBetween,
                             children: [
@@ -129,30 +133,32 @@ class WeaponDisplay extends ConsumerWidget {
                                 text: name,
                                 size: 16,
                               ),
-                              GestureDetector(
-                                onLongPress: () {
-                                  ref
-                                      .read(inventoryProvider.notifier)
-                                      .deleteWeaponFromInventory(
-                                          guid: guid, ref: ref);
-                                },
-                                child: Padding(
-                                  padding: const EdgeInsets.all(4.0),
-                                  child: AnimatedContainer(
-                                    duration: Duration(milliseconds: 200),
-                                    decoration: BoxDecoration(
-                                        color: Colors.transparent,
-                                        borderRadius: BorderRadius.circular(5),
-                                        border: Border.all(
-                                            width: 1, color: Colors.white)),
-                                    child: Icon(
-                                      Icons.delete_forever,
-                                      color: ref.watch(themeProvider).cardBg,
-                                      size: 35,
+                              if (canDelete)
+                                GestureDetector(
+                                  onLongPress: () {
+                                    ref
+                                        .read(inventoryProvider.notifier)
+                                        .deleteWeaponFromInventory(
+                                            guid: guid, ref: ref);
+                                  },
+                                  child: Padding(
+                                    padding: const EdgeInsets.all(4.0),
+                                    child: AnimatedContainer(
+                                      duration: Duration(milliseconds: 200),
+                                      decoration: BoxDecoration(
+                                          color: Colors.transparent,
+                                          borderRadius:
+                                              BorderRadius.circular(5),
+                                          border: Border.all(
+                                              width: 1, color: Colors.white)),
+                                      child: Icon(
+                                        Icons.delete_forever,
+                                        color: ref.watch(themeProvider).cardBg,
+                                        size: 35,
+                                      ),
                                     ),
                                   ),
-                                ),
-                              )
+                                )
                             ],
                           ),
                         ),
